+++ CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- First implementation which maps all 26 States and DF to a close timezone and 16 Cities to a specific timezone
+
 [Unreleased]: https://github.com/aryelgois/br-timezone/compare/initial-commit...HEAD